Add files via upload
Identical data to previous however with an additional column including information on the patients cancer stage, so that we can eliminate stages > T3b as the cancer has spread to the seminal vesicles and thus have been contoured also. Note that this information was not available for patients in the AllData19Frac.csv file, where the stage takes the value 'nil'.